refactor(router): split router middleware into separate files. Give it a dedicated directory.
